+++ importmap.php
@@ -65,8 +65,8 @@ return [
         'path' => './assets/js/pages/settings/index.js',
         'entrypoint' => true,
     ],
-    'js/pages/dechess/index.js' => [
-        'path' => './assets/js/pages/dechess/index.js',
+    'js/pages/connect/index.js' => [
+        'path' => './assets/js/pages/connect/index.js',
         'entrypoint' => true,
     ],
     'js/pages/edit/index.js' => [
