+++ src/Controller/Pages/Connect/ConnectController.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace App\Controller\Pages\DeChess;
+namespace App\Controller\Pages\Connect;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 
-class DeChessController extends AbstractController
+class ConnectController extends AbstractController
 {
     public function index(): Response
     {
-        return $this->render('pages/dechess/index.html.twig');
+        return $this->render('pages/connect/index.html.twig');
     }
 }
